Fix: Kit-Library - Code Review
ref: ED-2712
