+++ core/app/modules/kit-library/assets/js/data/kits/commands-data/download-link.js
@@ -1,4 +1,4 @@
-export class KitsDownloadLink extends $e.modules.CommandData {
+export class DownloadLink extends $e.modules.CommandData {
 	static getEndpointFormat() {
 		return 'kits/download-link/{id}';
 	}
