+++ core/app/modules/kit-library/assets/js/data/kits/component.js
@@ -1,0 +1,11 @@
+import * as commandsData from './commands-data/';
+
+export default class Component extends $e.modules.ComponentBase {
+	getNamespace() {
+		return 'kits';
+	}
+
+	defaultData() {
+		return this.importCommands( commandsData );
+	}
+}
