+++ core/app/modules/kit-library/data/kits/controller.php
@@ -1,23 +1,24 @@
 <?php
-namespace Elementor\Core\App\Modules\KitLibrary\Data;
+namespace Elementor\Core\App\Modules\KitLibrary\Data\Kits;
 
 use Elementor\Data\Base\Controller as Controller_Base;
+use Elementor\Core\App\Modules\KitLibrary\Data\Repository;
 use Elementor\Core\App\Modules\KitLibrary\Data\Exceptions\Wp_Error_Exception;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
 }
 
-class Kits_Controller extends Controller_Base {
+class Controller extends Controller_Base {
 	/**
 	 * @var Repository
 	 */
-	public $repository;
+	private $repository;
 
 	/**
 	 * @param \WP_REST_Request $request
 	 *
-	 * @return \WP_Error|\WP_REST_Response
+	 * @return \WP_Error|array
 	 */
 	public function get_items( $request ) {
 		try {
@@ -28,15 +29,15 @@ class Kits_Controller extends Controller_Base {
 			return new \WP_Error( 'server_error', __( 'Something went wrong.', 'elementor' ) );
 		}
 
-		return new \WP_REST_Response( [
+		return [
 			'data' => $data->all(),
-		] );
+		];
 	}
 
 	/**
 	 * @param \WP_REST_Request $request
 	 *
-	 * @return \WP_Error|\WP_REST_Response
+	 * @return \WP_Error|array
 	 */
 	public function get_item( $request ) {
 		try {
@@ -51,9 +52,9 @@ class Kits_Controller extends Controller_Base {
 			return new \WP_Error( 'not_found', __( 'Kit not found.', 'elementor' ), [ 'status' => 404 ] );
 		}
 
-		return new \WP_REST_Response( [
+		return [
 			'data' => $data,
-		] );
+		];
 	}
 
 	/**
@@ -67,7 +68,14 @@ class Kits_Controller extends Controller_Base {
 	 * Must implement.
 	 */
 	public function register_endpoints() {
-		$this->register_endpoint( Endpoints\Kits_Download_Link::class );
+		$this->register_endpoint( Endpoints\Download_Link::class );
+	}
+
+	/**
+	 * @return Repository
+	 */
+	public function get_repository() {
+		return $this->repository;
 	}
 
 	/**
@@ -75,7 +83,7 @@ class Kits_Controller extends Controller_Base {
 	 */
 	protected function register_internal_endpoints() {
 		// Register as internal to remove the default endpoint generated by the base controller.
-		$this->register_endpoint( Endpoints\Kits_Index::class );
+		$this->register_endpoint( Endpoints\Index::class );
 	}
 
 	/**
