+++ core/app/modules/kit-library/data/kits/endpoints/index.php
@@ -1,5 +1,5 @@
 <?php
-namespace Elementor\Core\App\Modules\KitLibrary\Data\Endpoints;
+namespace Elementor\Core\App\Modules\KitLibrary\Data\Kits\Endpoints;
 
 use Elementor\Data\Base\Endpoint;
 
@@ -7,7 +7,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
 }
 
-class Taxonomies_Index extends Endpoint {
+class Index extends Endpoint {
 	public function get_name() {
 		return 'index';
 	}
@@ -25,6 +25,21 @@ class Taxonomies_Index extends Endpoint {
 					'required' => false,
 					'default' => false,
 					'type' => 'boolean',
+				],
+			]
+		);
+
+		$this->register_route(
+			'(?P<id>[\w-]+)/',
+			\WP_REST_Server::READABLE,
+			function ( $request ) {
+				return $this->base_callback( \WP_REST_Server::READABLE, $request, false );
+			},
+			[
+				'id' => [
+					'description' => 'Unique identifier for the object.',
+					'type' => 'string',
+					'required' => true,
 				],
 			]
 		);
