+++ core/app/modules/kit-library/data/taxonomies/controller.php
@@ -1,14 +1,15 @@
 <?php
-namespace Elementor\Core\App\Modules\KitLibrary\Data;
+namespace Elementor\Core\App\Modules\KitLibrary\Data\Taxonomies;
 
 use Elementor\Data\Base\Controller as Controller_Base;
+use Elementor\Core\App\Modules\KitLibrary\Data\Repository;
 use Elementor\Core\App\Modules\KitLibrary\Data\Exceptions\Wp_Error_Exception;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
 }
 
-class Taxonomies_Controller extends Controller_Base {
+class Controller extends Controller_Base {
 	/**
 	 * @var Repository
 	 */
@@ -27,13 +28,13 @@ class Taxonomies_Controller extends Controller_Base {
 	 */
 	protected function register_internal_endpoints() {
 		// Register as internal to remove the default endpoint generated by the base controller.
-		$this->register_endpoint( Endpoints\Taxonomies_Index::class );
+		$this->register_endpoint( Endpoints\Index::class );
 	}
 
 	/**
 	 * @param \WP_REST_Request $request
 	 *
-	 * @return \WP_Error|\WP_REST_Response
+	 * @return \WP_Error|array
 	 */
 	public function get_items( $request ) {
 		try {
@@ -44,9 +45,9 @@ class Taxonomies_Controller extends Controller_Base {
 			return new \WP_Error( 'server_error', __( 'Something went wrong.', 'elementor' ) );
 		}
 
-		return new \WP_REST_Response( [
+		return [
 			'data' => $data->all(),
-		] );
+		];
 	}
 
 	/**
